filter split in to 2 separate components, show filter only for reviewer
